+++ POS/POS/Modules/Payout/ViewModels/Prompts/CashDrawerHistoryPromptViewModel.cs
@@ -1,4 +1,6 @@
-﻿using Framework.WPF.ScreenManagement.Prompt;
+﻿using System;
+using System.Collections.Generic;
+using Framework.WPF.ScreenManagement.Prompt;
 using POS.Core;
 using POS.Core.Interfaces.Data;
 using POS.Core.Session;
@@ -6,6 +8,14 @@ using POS.Modules.CashDrawerHistorys.Models;
 using POS.Modules.Main.ViewModels;
 using POS.Modules.Payout.Models;
 using System.Collections.ObjectModel;
+using System.Linq;
+using Framework.Core.Logging;
+using Framework.Infrastructure.Identity.Services;
+using Framework.WPF.UI.Controls;
+using POS.Common;
+using POS.Core.CashDrawer;
+using POS.Core.Interfaces.Printer;
+using POS.Core.Transaction;
 
 namespace POS.Modules.Payout.ViewModels.Prompts
 {
@@ -13,6 +23,10 @@ namespace POS.Modules.Payout.ViewModels.Prompts
     {
         private readonly Session _session;
         private readonly ICashDrawerRepository _cashDrawerRepository;
+        private readonly IPrintService _printService;
+        private readonly ILogEventDataService _logEventDataService;
+        private readonly IUserSession _userSession;
+        private readonly IMessageBoxService _messageBoxService;
 
         #region NotifyChangeProperties
         public ObservableCollection<CashDrawerHistory> _cashDrawerHistory;
@@ -29,11 +43,14 @@ namespace POS.Modules.Payout.ViewModels.Prompts
         }
         #endregion
 
-        public CashDrawerHistoryPromptViewModel(Session session, ICashDrawerRepository cashDrawerRepository)
+        public CashDrawerHistoryPromptViewModel(Session session, ICashDrawerRepository cashDrawerRepository, IPrintService printService, ILogEventDataService logEventDataService, IUserSession userSession, IMessageBoxService messageBoxService)
         {            
             _session = session;
             _cashDrawerRepository = cashDrawerRepository;
-
+            _printService = printService;
+            _logEventDataService = logEventDataService;
+            _userSession = userSession;
+            _messageBoxService = messageBoxService;
             Initialize();
         }
 
@@ -47,6 +64,37 @@ namespace POS.Modules.Payout.ViewModels.Prompts
                 (
                     CashDrawerHistoryTranslator.Translate(cashDrawerHistory)
                 );
+        }
+
+        public async void PrintHistory()
+        {
+            try
+            {
+                var printList = new List<(string Type, Double Amount, DateTime TransactionDate)>();
+                CashDrawerHistory.ToList().ForEach(item =>
+                {
+                    printList.Add(
+                        (item.TypeName, item.Amount, item.CreatedDate)
+                    );
+                });
+                var printHistory = new PrintCashDrawerHistory(_session.Username, printList);
+                _printService.PrintCashHistoryReceipt(printHistory);
+
+                await _logEventDataService.LogEventToDatabaseAsync(CashDrawerHistoryEventType.CashDrawerHistoryPrintSuccess,CashDrawerHistoryEventType.CashDrawerHistoryPrintSuccess.ToString(),
+                    $"SessionId: {_session.Id.Value}", _userSession.UserId);
+                await _messageBoxService.PromptAsync(POSResources.CashHistoryPrintedSuccessfully, POSResources.SuccessTitle, PromptOptions.Ok, PromptTypes.Success);
+                await base.Yes();
+            }
+
+            catch (Exception exception)
+            {
+                await _logEventDataService.LogEventToDatabaseAsync(CashDrawerHistoryEventType.CashDrawerHistoryPrintFail, CashDrawerHistoryEventType.CashDrawerHistoryPrintFail.ToString() + " " + exception.Message, exception.ToString(), _userSession.UserId);
+                await _messageBoxService.PromptAsync(POSResources.CashHistoryPrintFailed, POSResources.ErrorTitle, PromptOptions.Ok, PromptTypes.Error);
+                await base.Yes();
+                //await HandleErrorAsync(message + Environment.NewLine + exception.Message, exception);
+            }
+
+
         }
     }
 }